+++ src/main/java/net/medievalweapons/mixin/BowItemMixin.java
@@ -24,23 +24,23 @@ public abstract class BowItemMixin extends RangedWeaponItem {
 
     @Inject(method = "shoot", at = @At("TAIL"))
     private void onStoppedUsingMixin(LivingEntity shooter, ProjectileEntity projectile, int index, float speed, float divergence, float yaw, @Nullable LivingEntity target, CallbackInfo info) {
-        if (((BowItem) (Object) this) == ItemInit.LONG_BOW_ITEM) {
-            projectile.setVelocity(shooter, shooter.getPitch(), shooter.getYaw(), 0.0F, speed * 4.5F, divergence);
-        } else if (((BowItem) (Object) this) == ItemInit.RECURVE_BOW_ITEM) {
-            projectile.setVelocity(shooter, shooter.getPitch(), shooter.getYaw(), 0.0F, speed * 2.0F, divergence);
+        if (this == ItemInit.LONG_BOW_ITEM) {
+            projectile.setVelocity(shooter, shooter.getPitch(), shooter.getYaw(), 0.0F, speed * 1.3F, divergence);
+        } else if (this == ItemInit.RECURVE_BOW_ITEM) {
+            projectile.setVelocity(shooter, shooter.getPitch(), shooter.getYaw(), 0.0F, speed * 0.9F, divergence);
         }
     }
 
     @ModifyVariable(method = "onStoppedUsing", at = @At(value = "INVOKE_ASSIGN", target = "Lnet/minecraft/item/BowItem;getPullProgress(I)F"), ordinal = 0)
     private float onStoppedUsingPullProgressMixin(float original, ItemStack stack, World world, LivingEntity user, int remainingUseTicks) {
-        if (((BowItem) (Object) this) == ItemInit.LONG_BOW_ITEM) {
+        if (this == ItemInit.LONG_BOW_ITEM) {
             float f = (float) (this.getMaxUseTime(stack, user) - remainingUseTicks) / 60.0F;
             f = (f * f + f * 2.0F) / 3.0F;
             if (f > 1.0F) {
                 f = 1.0F;
             }
             return f;
-        } else if (((BowItem) (Object) this) == ItemInit.RECURVE_BOW_ITEM) {
+        } else if (this == ItemInit.RECURVE_BOW_ITEM) {
             float f = (float) (this.getMaxUseTime(stack, user) - remainingUseTicks) / 12.0F;
             f = (f * f + f * 2.0F / 3.0F);
             if (f > 1.0F) {